+++ app/Filament/Resources/PretixConnectionResource/Pages/CreatePretixConnection.php
@@ -1,0 +1,11 @@
+<?php
+
+namespace App\Filament\Resources\PretixConnectionResource\Pages;
+
+use App\Filament\Resources\PretixConnectionResource;
+use Filament\Resources\Pages\CreateRecord;
+
+class CreatePretixConnection extends CreateRecord
+{
+    protected static string $resource = PretixConnectionResource::class;
+}